Final extracting reviews and sentimental analysis
Sentiment Analysis of Hotel Reviews is an NLP based project whose main aim is to deal with the reviews of the user and predict its sentiment as Positive or Negative. Like many reviews for a particular hotel written by users on the website about the quality of hotel service, food, surrounding, etc, then the role of our model is to predict the sentiment of those reviews, which help users in finding a better quality of the hotel according to the analysis of the review in future.