String find() Method in Python
Python find() method is used to find the index of the first occurrence of a sub-string present in the given string within an index range.

In other words, the find() method will search for a sub-string in the given string and returns the starting index of the sub-string if found. Otherwise, this method will return -1.