Palindrome a String
A Palindrome String is a collection of alphabets that remains the exact same way when read backward. They are also called Symmetrical Alphabets. When its alphabets are written in reverse order, they turn out to be the exact same combination of alphabets as the original string. For eg: “madam” is a Palindrome. its alphabets are reversed, it again becomes “madam” which was our original string. “napkin” is not a Palindrome. When reversed, the new number becomes “nikpan” which is different from the original string.